+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Renamed SshBasedConnection to Connection
+- Connection now accepts any instance of Transport
 
 ## 0.1.0 - 2020-03-26
 ### Added